requirements.txt
flask==3.1.0
ollama==0.4.7
python-dotenv==1.0.1
